Updated logic for local planner to respond to changes in state.  Added abort task script.

+++ src/acfr/global-planner/global_planner.cpp
@@ -197,7 +197,7 @@ int GlobalPlanner::clock() {
 		}
 		
 		// Send the global state change message
-		cout << "Publishing new global state: " << gpState.state << endl;
+		cout << "Publishing new global state: " << (int)(gpState.state) << endl;
 		lcm.publish("GLOBAL_STATE", &gpState);
 	}
 	globalPlannerMessage = globalPlannerIdle;
